+++ app/src/main/java/com/amrdeveloper/codeviewlibrary/syntax/LanguageName.java
@@ -3,7 +3,7 @@ package com.amrdeveloper.codeviewlibrary.syntax;
 /**
  * List of examples languages for CodeView
  */
-public enum Language {
+public enum LanguageName {
     JAVA,
     PYTHON,
     GO_LANG