+++ codeview/src/main/java/com/amrdeveloper/codeview/CodeView.java
@@ -25,6 +25,7 @@
 package com.amrdeveloper.codeview;
 
 import android.content.Context;
+import android.content.res.Resources;
 import android.graphics.Canvas;
 import android.graphics.Color;
 import android.graphics.Paint;
@@ -44,6 +45,8 @@ import android.text.style.CharacterStyle;
 import android.text.style.ForegroundColorSpan;
 import android.text.style.ReplacementSpan;
 import android.util.AttributeSet;
+import android.view.KeyEvent;
+import android.view.View;
 import android.widget.MultiAutoCompleteTextView;
 
 import androidx.annotation.ColorInt;
@@ -54,7 +57,6 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Locale;
 import java.util.Map;
 import java.util.Set;
 import java.util.SortedMap;
@@ -93,6 +95,12 @@ public class CodeView extends AppCompatMultiAutoCompleteTextView implements Find
     private CharacterStyle currentMatchedToken;
     private final List<Token> matchedTokens = new ArrayList<>();
 
+    private int maxNumberOfSuggestions = Integer.MAX_VALUE;
+    private int autoCompleteItemHeightInDp = (int) (50 * Resources.getSystem().getDisplayMetrics().density);
+
+    private boolean enablePairComplete = false;
+    private final Map<Character, Character> mPairCompleteMap = new HashMap<>();
+
     private final Handler mUpdateHandler = new Handler();
     private MultiAutoCompleteTextView.Tokenizer mAutoCompleteTokenizer;
 
@@ -125,6 +133,7 @@ public class CodeView extends AppCompatMultiAutoCompleteTextView implements Find
         setHorizontallyScrolling(true);
         setFilters(new InputFilter[]{mInputFilter});
         addTextChangedListener(mEditorTextWatcher);
+        setOnKeyListener(mOnKeyListener);
 
         lineNumberRect = new Rect();
         lineNumberPaint = new Paint(Paint.ANTI_ALIAS_FLAG);
@@ -134,15 +143,17 @@ public class CodeView extends AppCompatMultiAutoCompleteTextView implements Find
     @Override
     protected void onDraw(Canvas canvas) {
         if (enableLineNumber) {
+            final Editable fullText = getText();
+            final Layout layout = getLayout();
+            final int lineCount = getLineCount();
             int baseline;
-            int lineCount = getLineCount();
-            int lineNumber = 1;
+            int currentLineNumber = 1;
 
             for (int i = 0; i < lineCount; ++i) {
                 baseline = getLineBounds(i, null);
-                if (i == 0 || getText().charAt(getLayout().getLineStart(i) - 1) == '\n') {
-                    canvas.drawText(String.format(Locale.ENGLISH, " %d", lineNumber), lineNumberRect.left, baseline, lineNumberPaint);
-                    ++lineNumber;
+                if (i == 0 || fullText.charAt(layout.getLineStart(i) - 1) == '\n') {
+                    canvas.drawText(" " + currentLineNumber, lineNumberRect.left, baseline, lineNumberPaint);
+                    ++currentLineNumber;
                 }
             }
 
@@ -578,22 +589,103 @@ public class CodeView extends AppCompatMultiAutoCompleteTextView implements Find
     lineNumberPaint.setTypeface(typeface);
     }
 
+    /**
+     * Modify the maximum number of suggestions to show, default is Integer.MAX_VALUE
+     * @param maxSuggestions the maximum number of suggestions
+     * @since 1.3.0
+     */
+    public void setMaxSuggestionsSize(int maxSuggestions) {
+        maxNumberOfSuggestions = maxSuggestions;
+    }
+
+    /**
+     * Modify the auto complete item height
+     * @param height auto complete item height in dp
+     * @since 1.3.0
+     */
+    public void setAutoCompleteItemHeightInDp(int height) {
+        autoCompleteItemHeightInDp = (int) (height * Resources.getSystem().getDisplayMetrics().density);
+    }
+
+    /**
+     * Enable or disable the auto pairs complete feature
+     * @param enable Flag to enable or disable auto pair complete
+     * @since 1.3.0
+     */
+    public void enablePairComplete(boolean enable) {
+        enablePairComplete = enable;
+    }
+
+    /**
+     * Set the pairs for auto pairs complete feature
+     * @param map Map of pairs of characters
+     * @since 1.3.0
+     */
+    public void setPairCompleteMap(Map<Character, Character> map) {
+        mPairCompleteMap.clear();
+        mPairCompleteMap.putAll(map);
+    }
+
+    /**
+     * Add new pair complete item using key and value
+     * @param key the pair complete item key
+     * @param value the pair complete item value
+     * @since 1.3.0
+     */
+    public void addPairCompleteItem(char key, char value) {
+        mPairCompleteMap.put(key, value);
+    }
+
+    /**
+     * Remove single pair complete item by key
+     * @param key the pair complete item key
+     * @since 1.3.0
+     */
+    public void removePairCompleteItem(char key) {
+        mPairCompleteMap.remove(key);
+    }
+
+    /**
+     * Clear all of pairs
+     * @since 1.3.0
+     */
+    public void clearPairCompleteMap() {
+        mPairCompleteMap.clear();
+    }
+
     @Override
     public void showDropDown() {
-        int[] screenPoint = new int[2];
+        final int[] screenPoint = new int[2];
         getLocationOnScreen(screenPoint);
+
+        final Layout layout = getLayout();
+        final int position = getSelectionStart();
+        final int line = layout.getLineForOffset(position);
+        final int lineButton = layout.getLineBottom(line);
+
+        int numberOfMatchedItems = getAdapter().getCount();
+        if (numberOfMatchedItems > maxNumberOfSuggestions) {
+            numberOfMatchedItems = maxNumberOfSuggestions;
+        }
+
+        int dropDownHeight = getDropDownHeight();
+        int modifiedDropDownHeight = numberOfMatchedItems * autoCompleteItemHeightInDp;
+        if (dropDownHeight != modifiedDropDownHeight) {
+            dropDownHeight = modifiedDropDownHeight;
+        }
 
         final Rect displayFrame = new Rect();
         getWindowVisibleDisplayFrame(displayFrame);
+        int displayFrameHeight = displayFrame.height();
 
-        int position = getSelectionStart();
-        Layout layout = getLayout();
-        int line = layout.getLineForOffset(position);
-        int lineButton = layout.getLineBottom(line);
-        int dropDownVerticalOffset = lineButton + 140;
-        int dropDownHorizontalOffset = (int) layout.getPrimaryHorizontal(position);
-        setDropDownVerticalOffset(dropDownVerticalOffset);
-        setDropDownHorizontalOffset(dropDownHorizontalOffset);
+        int verticalOffset = lineButton + dropDownHeight;
+        if (verticalOffset > displayFrameHeight) {
+            verticalOffset = displayFrameHeight - autoCompleteItemHeightInDp;
+        }
+
+        setDropDownHeight(dropDownHeight);
+        setDropDownVerticalOffset(verticalOffset);
+        setDropDownHorizontalOffset((int) layout.getPrimaryHorizontal(position));
 
         super.showDropDown();
     }
@@ -603,6 +695,26 @@ public class CodeView extends AppCompatMultiAutoCompleteTextView implements Find
         public void run() {
             Editable source = getText();
             highlightWithoutChange(source);
+        }
+    };
+
+    private final OnKeyListener mOnKeyListener = new OnKeyListener() {
+
+        @Override
+        public boolean onKey(View v, int keyCode, KeyEvent event) {
+            if (!enableAutoIndentation) return false;
+            if (event.getAction() != KeyEvent.ACTION_DOWN)
+                return true;
+            switch (keyCode) {
+                case KeyEvent.KEYCODE_SPACE:
+                    currentIndentation++;
+                    break;
+                case KeyEvent.KEYCODE_DEL:
+                    if (currentIndentation > 0)
+                        currentIndentation--;
+                    break;
+            }
+            return false;
         }
     };
 
@@ -630,11 +742,30 @@ public class CodeView extends AppCompatMultiAutoCompleteTextView implements Find
 
             if (mRemoveErrorsWhenTextChanged) removeAllErrorLines();
 
-            if (enableAutoIndentation && count == 1) {
-                if (indentationStarts.contains(charSequence.charAt(start)))
-                    currentIndentation += tabLength;
-                else if (indentationEnds.contains(charSequence.charAt(start)))
-                    currentIndentation -= tabLength;
+            if (count == 1 && (enableAutoIndentation || enablePairComplete)) {
+                char currentChar = charSequence.charAt(start);
+
+                if (enableAutoIndentation) {
+                    if (indentationStarts.contains(currentChar))
+                        currentIndentation += tabLength;
+                    else if (indentationEnds.contains(currentChar))
+                        currentIndentation -= tabLength;
+                }
+
+                if (enablePairComplete) {
+                    Character pairValue = mPairCompleteMap.get(currentChar);
+                    if (pairValue != null) {
+                        modified = false;
+                        getText().insert(getSelectionEnd(), pairValue.toString());
+                        if (enableAutoIndentation) {
+                            if (indentationStarts.contains(pairValue))
+                                currentIndentation += tabLength;
+                            else if (indentationEnds.contains(pairValue))
+                                currentIndentation -= tabLength;
+                        }
+                        modified = true;
+                    }
+                }
             }
         }
 
@@ -674,11 +805,17 @@ public class CodeView extends AppCompatMultiAutoCompleteTextView implements Find
         public CharSequence filter(CharSequence source, int start, int end,
                                    Spanned dest, int dStart, int dEnd) {
             if (modified && enableAutoIndentation && start < source.length()) {
-                boolean isInsertedAtEnd = dest.length() == dEnd;
                 if (source.charAt(start) == '\n') {
-                    int indentation = isInsertedAtEnd
-                            ? currentIndentation
-                            : calculateSourceIndentation(dest.subSequence(0, dStart));
+                    // Apply the current indentation if it inserted at the end
+                    if (dest.length() == dEnd) return applyIndentation(source, currentIndentation);
+
+                    // reCalculate the current indentation
+                    int indentation = calculateSourceIndentation(dest.subSequence(0, dStart));
+
+                    // Decrement the indentation if the next char is on indentationEnds set
+                    if (indentationEnds.contains(dest.charAt(dEnd))) indentation -= tabLength;
+
+                    // Apply the new indentation to the source code
                     return applyIndentation(source, indentation);
                 }
             }